make new component with reusable files, change name row, change 2 components into 1
